app/requirements.txt
python-dotenv==1.0.0
requests==2.31.0
boto3==1.26.142